% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R, R/zzz.R
\docType{package}
\name{ardis}
\alias{ardis}
\alias{ardis-package}
\title{Create a ardis table object}
\usage{
ardis(target, treat_var, where = TRUE, cols = vars())
}
\arguments{
\item{target}{Dataset upon which summaries will be performed}

\item{treat_var}{Variable containing treatment group assignments. Supply unquoted.}

\item{where}{A general subset to be applied to all layers. Supply as programming logic (i.e. x < 5 & y == 10)}

\item{cols}{A grouping variable to summarize data by column (in addition to treat_var). Provide multiple
column variables by using \code{\link[dplyr]{vars}}}
}
\value{
A \code{ardis} object
}
\description{
The \code{ardis} object is the main container upon which a ardis table is constructed. ardis tables are made up of
one or more layers. Each layer contains an instruction for a summary to be performed. The \code{ardis} object contains
those layers, and the general data, metadata, and logic necessary.

`r lifecycle::badge("experimental")`
}
\details{
When a \code{ardis} is created, it will contain the following bindings:
\itemize{
\item{target - The dataset upon which summaries will be performed}
\item{pop_data - The data containing population information. This defaults to the target dataset}
\item{cols - A categorical variable to present summaries grouped by column (in addition to treat_var)}
\item{table_where - The \code{where} parameter provided, used to subset the target data}
\item{treat_var - Variable used to distinguish treatment groups.}
\item{header_n - Default header N values based on \code{treat_var}}
\item{pop_treat_var - The treatment variable for \code{pop_data} (if different)}
\item{layers - The container for individual layers of a \code{ardis}}
\item{treat_grps - Additional treatment groups to be added to the summary (i.e. Total)}
}

\code{ardis} allows you a basic interface to instantiate the object. Modifier functions are available to change
individual parameters catered to your analysis. For example, to add a total group, you can use the
\code{\link{add_total_group}}.

In future releases, we will provide vignettes to fully demonstrate these capabilities.

'ardis' is a package dedicated to simplifying the data manipulation necessary
to create clinical reports. Clinical data summaries can often be broken down
into two factors - counting discrete variables (or counting shifts in state),
and descriptive statistics around a continuous variable. Many of the reports
that go into a clinical report are made up of these two scenarios. By
abstracting this process away, 'ardis' allows you to rapidly build these
tables without worrying about the underlying data manipulation.

'ardis' takes this process a few steps further by abstracting away most of
the programming that goes into proper presentation, which is where a great
deal of programming time is spent. For example, 'ardis' allows you to easily
control:

\describe{ \item{\strong{String formatting}}{Different reports warrant
different presentation of your strings. Programming this can get tedious, as
you typically want to make sure that your decimals properly align. 'ardis'
abstracts this process away and provides you with a simple interface to
specify how you want your data presented} \item{\strong{Treatment
groups}}{Need a total column? Need to group summaries of multiple treatments?
'ardis' makes it simple to add additional treatment groups into your report}
\item{\strong{Denominators}}{n (\%) counts often vary based on the summary
being performed. 'ardis' allows you to easily control what denominators are
used based on a few common scenarios} \item{\strong{Sorting}}{Summarizing
data is one thing, but ordering it for presentation. ardis automatically
derives sorting variable to give you the data you need to order your table
properly. This process is flexible so you can easily get what you want by
leveraging your data or characteristics of R.} }

Another powerful aspect of 'ardis' are the objects themselves. 'ardis' does
more than format your data. Metadata about your table is kept under the hood,
and functions allow you to access information that you need. For example,
'ardis' allows you to calculate and access the raw numeric data of
calculations as well, and easily pick out just the pieces of information that
you need.

Lastly, 'ardis' was built to be flexible, yet intuitive. A common pitfall of
building tools like this is over automation. By doing to much, you end up not
doing enough. 'ardis' aims to hit the sweet spot in between. Additionally, we
designed our function interfaces to be clean. Modifier functions offer you
flexibility when you need it, but defaults can be set to keep the code
concise. This allows you to quickly assemble your table, and easily make
changes where necessary.
}
\examples{

tab <- ardis(iris, Species, where = Sepal.Length < 5.8)

# Load in pipe
library(magrittr)

# Use just the defaults
ardis(mtcars, gear) \%>\%
  add_layer(
    group_desc(mpg, by=cyl)
  ) \%>\%
  add_layer(
    group_count(carb, by=cyl)
  ) \%>\%
  build()

# Customize and modify
ardis(mtcars, gear) \%>\%
  add_layer(
    group_desc(mpg, by=cyl) \%>\%
      set_format_strings(
        "n"         = f_str("xx", n),
        "Mean (SD)" = f_str("a.a+1 (a.a+2)", mean, sd, empty='NA'),
        "Median"    = f_str("a.a+1", median),
        "Q1, Q3"    = f_str("a, a", q1, q3, empty=c(.overall='NA')),
        "Min, Max"  = f_str("a, a", min, max),
        "Missing"   = f_str("xx", missing)
      )
  ) \%>\%
  add_layer(
    group_count(carb, by=cyl) \%>\%
      add_risk_diff(
        c('5', '3'),
        c('4', '3')
      ) \%>\%
      set_format_strings(
        n_counts = f_str('xx (xx\%)', n, pct),
        riskdiff = f_str('xx.xxx (xx.xxx, xx.xxx)', dif, low, high)
      ) \%>\%
      set_order_count_method("bycount") \%>\%
      set_ordering_cols('4') \%>\%
      set_result_order_var(pct)
  ) \%>\%
  build()

# A Shift Table
ardis(mtcars, am) \%>\%
  add_layer(
    group_shift(vars(row=gear, column=carb), by=cyl) \%>\%
    set_format_strings(f_str("xxx (xx.xx\%)", n, pct))
  ) \%>\%
  build()

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atorus-research/ardis}
  \item Report bugs at \url{https://github.com/atorus-research/ardis/issues}
}

}
\author{
\strong{Maintainer}: Mike Stackhouse \email{mike.stackhouse@atorusresearch.com} (\href{https://orcid.org/0000-0001-6030-723X}{ORCID})

Authors:
\itemize{
  \item Eli Miller \email{Eli.Miller@AtorusResearch.com} (\href{https://orcid.org/0000-0002-2127-9456}{ORCID})
}

Other contributors:
\itemize{
  \item Atorus Research LLC [copyright holder]
}

}
