% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Retrieve the numeric data from a ardis objects}
\usage{
build(x, layer = NULL, where = TRUE, ...)
}
\arguments{
\item{x}{A ardis or ardis_layer object}

\item{layer}{Layer name or index to select out specifically}

\item{where}{Subset criteria passed to dplyr::filter}

\item{...}{Additional arguments to pass forward}
}
\value{
Numeric data from the ardis layer
}
\description{
\code{build} provides access to the un-formatted numeric data for
each of the layers within a \code{ardis}, with options to allow you to
extract distinct layers and filter as desired.
}
\details{
When used on a \code{ardis} object, this method will aggregate the
numeric data from all ardis layers. The data will be returned to the user in
a list of data frames. If the data has already been processed (i.e.
\code{build} has been run), the numeric data is already available and will be
returned without reprocessing. Otherwise, the numeric portion of the layer
will be processed.

Using the layer and where parameters, data for a specific layer can be
extracted and subset. This is most clear when layers are given text names
instead of using a layer index, but a numeric index works as well.
}
\examples{

t <- ardis(mtcars, gear) \%>\%
 add_layer(name='drat',
           group_desc(drat)
 ) \%>\%
 add_layer(name='cyl',
           group_count(cyl)
 )

 # Return a list of the numeric data frames
 build(t)

 # Get the data from a specific layer
 build(t, layer='drat')
 build(t, layer=1)

 # Choose multiple layers by name or index
 build(t, layer=c('cyl', 'drat'))
 build(t, layer=c(2, 1))

 # Get the data and filter it
 build(t, layer='drat', where = gear==3)

}
