% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_summaries.R
\name{has_summaries}
\alias{has_summaries}
\title{Check if summaries have been set on a ardis layer}
\usage{
has_summaries(e)
}
\arguments{
\item{e}{A layer environment}
}
\value{
Boolean
}
\description{
Check if summaries have been set on a ardis layer
}
