% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{process_statistic_data}
\alias{process_statistic_data}
\title{Process a ardis_statistic object}
\usage{
process_statistic_data(x, ...)
}
\arguments{
\item{x}{A ardis_statistic environment}

\item{...}{Additional pass through parameters}
}
\value{
Numeric statistc data from a ardis statistc
}
\description{
This is an internal function that is not meant for use externally, but must be exported.
Use with caution.
}
\keyword{internal}
