% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bindings.R
\name{get_desc_layer_summaries}
\alias{get_desc_layer_summaries}
\alias{set_desc_layer_summaries}
\alias{get_count_layer_summaries}
\alias{set_count_layer_summaries}
\alias{get_shift_layer_formats}
\alias{set_shift_layer_formats}
\title{Get or set the default format strings for descriptive statistics layers}
\usage{
get_desc_layer_summaries(obj)

set_desc_layer_summaries(obj, ...)

get_count_layer_summaries(obj)

set_count_layer_summaries(obj, ...)

get_shift_layer_formats(obj)

set_shift_layer_formats(obj, ...)
}
\arguments{
\item{obj}{A ardis object}

\item{...}{formats to pass forward}
}
\description{
ardis provides you with the ability to set table-wide defaults of format
strings. You may wish to reuse the same summaries across numerous
layers. \code{set_desc_layer_summaries} and \code{set_count_layer_summaries}
allow you to apply your desired format strings within the entire scope of the
table.
}
\details{
For descriptive statistic layers, you can also use \code{set_summaries}
and \code{set_desc_layer_summaries} together within a table, but not within the
same layer. In the absence of specified summaries, first the table will
be checked for any available defaults, and otherwise the
\code{ardis.desc_layer_default_formats} option will be used.
\code{set_summaries} will always take precedence over either. Defaults
cannot be combined between \code{set_summaries},
\code{set_desc_layer_summaries}, and the
\code{ardis.desc_layer_default_summaries} because the order of presentation of
results is controlled by the format strings, so relying on combinations of
these setting would not be intuitive.
}
