% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bindings.R
\name{treat_var}
\alias{treat_var}
\alias{set_treat_var}
\title{Return or set the treatment variable binding}
\usage{
treat_var(table)

set_treat_var(table, treat_var)
}
\arguments{
\item{table}{A \code{tardis_table} object to set or return treatment variable
the table is split by.}

\item{treat_var}{Variable containing treatment group assignments. Supply unquoted.}
}
\value{
For \code{tardis_treat_var} the treat_var binding of the \code{tardis_table}
  object. For \code{set_tardis_treat_var} the modified object.
}
\description{
Return or set the treatment variable binding
}
\examples{
tab <- tardis_table(mtcars, cyl)

set_treat_var(tab, gear)

}
