% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{get_metadata}
\alias{get_metadata}
\title{Get the metadata dataframe from a tardis_table}
\usage{
get_metadata(t)
}
\arguments{
\item{t}{A tardis table with metadata built}
}
\value{
tardis metadata dataframe
}
\description{
Pull out the metadata dataframe from a tardis_table to work with it directly
}
\examples{
t <- tardis_table(mtcars, gear) \%>\%
  add_layer(
    group_desc(wt)
  )

t \%>\%
  build(metadata=TRUE)

get_metadata(t)
}
