% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{add_variables}
\alias{add_variables}
\alias{add_filters}
\title{Add variables to a ardis_meta object}
\usage{
add_variables(meta, names)

add_filters(meta, filters)
}
\arguments{
\item{meta}{A ardis_meta object}

\item{names}{A list of names, providing variable names of interest. Provide
as a list of quosures using \code{rlang::quos()}}

\item{filters}{A list of symbols, providing variable names of interest. Provide
as a list of quosures using `rlang::quos()`}
}
\value{
ardis_meta object
}
\description{
Add additional variable names to a \code{ardis_meta()} object.
}
\examples{

m <- ardis_meta()
m <- add_variables(m, rlang::quos(a, b, c))
m <- add_filters(m, rlang::quos(a==1, b==2, c==3))
m
}
\concept{Metadata additions}
